fix: show preview with vite
