+++ .gitignore
@@ -1,6 +1,4 @@
 /node_modules
 /dist
 
-preview/index.html
-
 .FIGMA_TOKEN
