+++ index.html
@@ -3,40 +3,46 @@
   <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=Edge">
-    <title>WARP Icons</title>
     <meta name="description" content="All WARP icons" />
     <link href="https://assets.finn.no/pkg/@warp-ds/css/v1/tokens/finn-no.css" rel="stylesheet" />
+    <link href="https://assets.finn.no/pkg/@warp-ds/css/v1/resets.css" rel="stylesheet" />
     <style>
-    .max-w-screen-xl {
-      max-width: 1280px;
-    }
+      html {
+        font-size: 62.5%;
+        margin: 0;
+      }
+      .max-w-screen-xl {
+        max-width: 1280px;
+      }
     </style>
+    <script type="module" src="./main.js"></script>
   </head>
-  <body class="bg-gray-100">
-    <div class="bg-white border-b border-gray-200">
+  <body class="s-bg-active">
+    <div class="s-bg border-b s-border">
       <div class="max-w-screen-xl mx-auto px-32 py-16 flex items-center justify-between">
-        <h1 class="h3 antialiased mb-0">WARP Icons</h1>
-        <a class="text-gray-500 hover:text-gray-700" href="https://github.com/warp-ds/icons">
+        <h1 class="h1 antialiased mb-0">WARP Icons</h1>
+        <a class="s-text-subtle" href="https://github.com/warp-ds/icons">
           <svg class="h-24 w-24" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor"><title>GitHub repository</title><path d="M10 0a10 10 0 0 0-3.16 19.49c.5.1.68-.22.68-.48l-.01-1.7c-2.78.6-3.37-1.34-3.37-1.34-.46-1.16-1.11-1.47-1.11-1.47-.9-.62.07-.6.07-.6 1 .07 1.53 1.03 1.53 1.03.9 1.52 2.34 1.08 2.91.83.1-.65.35-1.09.63-1.34-2.22-.25-4.55-1.11-4.55-4.94 0-1.1.39-1.99 1.03-2.69a3.6 3.6 0 0 1 .1-2.64s.84-.27 2.75 1.02a9.58 9.58 0 0 1 5 0c1.91-1.3 2.75-1.02 2.75-1.02.55 1.37.2 2.4.1 2.64.64.7 1.03 1.6 1.03 2.69 0 3.84-2.34 4.68-4.57 4.93.36.31.68.92.68 1.85l-.01 2.75c0 .26.18.58.69.48A10 10 0 0 0 10 0"></path></svg>
         </a>
       </div>
     </div>
 
     <main class="max-w-screen-xl mx-auto px-32">
+      {% block content %}
         {% for size, icons in iconsBySize %}
           <section class="my-64">
-            <h2 class="h4 antialiased mb-16">{{size}}</h2>
-            <div class="grid gap-32 bg-white py-32 px-16 rounded-8 justify-center text-center" style="grid-template-columns: repeat(auto-fill, 128px)">
+            <h2 class="h3 antialiased mb-16">{{size}}</h2>
+            <div class="grid gap-32 s-bg py-32 px-16 rounded-8 justify-center text-center" style="grid-template-columns: repeat(auto-fill, 128px)">
                 {% for icon in icons %}
                   <div class="flex flex-col items-center">
                     {{icon.svg | safe }}
-                    <div class="text-gray-500 text-12 mt-8 font-bold antialiased">{{ icon.name }}</div>
+                    <div class="s-text-subtle text-16 mt-8 font-bold antialiased">{{ icon.name }}</div>
                   </div>
                 {% endfor %}
             </div>
           </section>
         {% endfor %}
+      {% endblock %}
     </main>
-
   </body>
-</html>
+</html>
