+++ package.json
@@ -1,41 +1,47 @@
 {
-    "name": "@warp-ds/icons",
-    "type": "module",
-    "version": "0.0.1",
-    "description": "WARP's icon set",
-    "main": "index.js",
-    "files": [
-      "dist"
-    ],
-    "exports": {
-      ".": "./index.js"
-    },
-    "scripts": {
-      "import": "node ./scripts/import.js",
-      "build": "node ./scripts/build.js",
-      "render": "node preview/render.js",
-      "start": "npm run build && node preview/render.js && open ./preview/index.html"
-    },
-    "license": "Apache-2.0",
-    "devDependencies": {
-      "@sindresorhus/slugify": "^2.1.0",
-      "camelcase": "^7.0.0",
-      "chalk": "^5.0.1",
-      "esbuild": "^0.14.42",
-      "fs-extra": "^10.1.0",
-      "glob": "^8.0.3",
-      "happy-dom": "^6.0.4",
-      "js-yaml": "^4.1.0",
-      "jsdom": "^16.5.3",
-      "nanoid": "^4.0.0",
-      "node-fetch": "^3.2.9",
-      "nunjucks": "^3.2.2",
-      "ora": "^6.1.2",
-      "prompts": "^2.4.0",
-      "rimraf": "^3.0.2",
-      "svgo": "^3.0.0"
-    },
-    "publishConfig": {
-      "access": "public"
-    }
+  "name": "@warp-ds/icons",
+  "type": "module",
+  "version": "0.0.1",
+  "description": "WARP's icon set",
+  "main": "index.js",
+  "files": [
+    "dist"
+  ],
+  "exports": {
+    ".": "./index.js"
+  },
+  "scripts": {
+    "import": "node ./scripts/import.js",
+    "build": "node ./scripts/build.js",
+    "render": "node preview/render.js",
+    "dev": "vite",
+    "vite:build": "vite build",
+    "preview": "pnpm run build && vite build && vite preview"
+  },
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "@sindresorhus/slugify": "^2.1.0",
+    "@warp-ds/uno": "^1.0.0",
+    "camelcase": "^7.0.0",
+    "chalk": "^5.0.1",
+    "esbuild": "^0.14.42",
+    "fs-extra": "^10.1.0",
+    "glob": "^8.0.3",
+    "happy-dom": "^6.0.4",
+    "js-yaml": "^4.1.0",
+    "jsdom": "^16.5.3",
+    "nanoid": "^4.0.0",
+    "node-fetch": "^3.2.9",
+    "nunjucks": "^3.2.2",
+    "ora": "^6.1.2",
+    "prompts": "^2.4.0",
+    "rimraf": "^3.0.2",
+    "svgo": "^3.0.0",
+    "unocss": "^0.55.2",
+    "vite": "^4.4.9",
+    "vite-plugin-nunjucks": "^0.1.10"
+  },
+  "publishConfig": {
+    "access": "public"
   }
+}
