+++ scripts/build.js
@@ -10,7 +10,6 @@ import { getElement, getNameAndSize, pascalCase } from './util/helpers.js'
 import { readFileSync } from 'node:fs'
 import yaml from 'js-yaml'
 import { basedir } from "../index.js";
-
 const descriptionsFile = readFileSync(path.join(basedir, 'icon-descriptions.yml'))
 const descriptions = yaml.load(descriptionsFile)
